Remove empty file
